requirements.txt
boto3==1.4.7
botocore==1.7.16
configparser==3.5.0
docutils==0.14
feedparser==5.2.1
httplib2==0.10.3
jmespath==0.9.3
mock==2.0.0
oauth2==1.9.0.post1
pbr==3.1.1
python-dateutil==2.6.1
s3transfer==0.1.11
python-dateutil==2.6.1
six==1.11.0
psycopg2-binary==2.7.7

